requirements.txt
python-dotenv
pywin32
selenium
webdriver-manager
pystray
Pillow
